change name on method and move it
