+++ client/src/produkter/tabs/VideosTab.tsx
@@ -7,11 +7,7 @@ import { useErrorStore } from "utils/store/useErrorStore";
 import { HM_REGISTER_URL } from "environments";
 import { MoreMenu } from "felleskomponenter/MoreMenu";
 import ReactPlayer from "react-player";
-import {
-  getEditedProductDTOAddMedia,
-  getEditedProductDTORemoveMedia,
-  mapImagesAndPDFfromMedia,
-} from "utils/product-util";
+import { getEditedProductDTOAddMedia, mapImagesAndPDFfromMedia, removeFileFromProduct } from "utils/product-util";
 
 const VideoTab = ({
   products,
@@ -96,7 +92,7 @@ const VideoTab = ({
     }
 
     const productToUpdate: ProductRegistrationDTO = await res.json();
-    const editedProductDTO = getEditedProductDTORemoveMedia(productToUpdate, uri);
+    const editedProductDTO = removeFileFromProduct(productToUpdate, uri);
 
     res = await fetch(`${HM_REGISTER_URL()}/admreg/vendor/api/v1/product/registrations/${productToUpdate.id}`, {
       method: "PUT",
